Time: 59 ms (62.46%) | Memory: 180.8 MB (56.67%) - LeetSync

+++ 307-range-sum-query-mutable/range-sum-query-mutable.cpp
@@ -2,7 +2,7 @@ class NumArray {
 public:
     struct segtree {
         int size;
-        vector<long long> sums;
+        vector<int> sums;
         void init(int n) {
             size = 1;
             while (size < n) size *= 2;
